src/{basic, short, sym}.rs: check for, panic on invalid sym resolution
We'd like any attempt to resolve a symbol on the wrong interner to be
a compile error; since this isn't possible with present-day Rust in
the general case, the next best thing is to panic on such an attempt.
This behavior and the additional data that supports it is enabled only
when the crate is compiled in debug mode.